Implement conditional GitHub status update
Fix a flaw in the update logic that allowed ovewriting all legit
states with an error state, e.g. getting owener after it was deleted
results in an error and that error could be sent to GitHub.
This change ensures that only error and pending states are treated
as transient, and failure or success states are final.