Implement promview
This is a very simple proxy that can forward metrics from prom
service inside the prom namespace in each of the test clusters.
It enables Prometheus federation without exposing test cluster
metrics to the outside world or somehow otherwise, and avoids
having to have a remote storage solution.

It doesn't not prevent introduction of remote storage by any means,
however the test cluster doesn't need to be away of remote storage
and, most importantly, there is no need for auth. So remote storage
can be implemented within management cluster without having to have
full auth setup, and/or ensuring metrics are ingested with with
unique labels.