Fix regression in how test cluster client set is constructed
As of introduction of the nodepool dependency, job creation and
clusterrolebinding creation code was moved to `CNRMContainerNodePoolWatcher`.
However, the `GKEClientSetBuilder` assumed that it gets `ContainerCluster` object,
but now it was getting `ContainerNodePool`. Due to the lack of typechecks on
`Unstructured` this didn't cause a buildtime error.

This change moves clusterrolebinding logic back into `CNRMContainerClusterWatcher`,
it alos introduces a test and runtime type check.