Use defaults for `autoRepair` and `autoUpgrade`
These two features had been deliberately disable because they were
deemed to interfere with tests. GKE API no longer allows to disable
these features when using REGULAR release channel.

It maybe possible to disable this once cluster version would be set
statically, but that's not supported by the operator yet.

It's important to note that both of these features are concerting
the nodepool, not the control plane.

With regards to auto-updades, it maybe viable to define a maintenance
window that is outside of the expect test duration, but it's not a
very trivial solution to something that is currently only a hypothetical
problem.

With regards to auto-repairs, there is also no known practical issue
at present.